Added delete shop command
