+++ src/next-app/cypress/support/commands.ts
@@ -40,11 +40,15 @@
 //   }
 // }
 
+import apiUrl from "../../constants/api-url";
+import Cookies from "js-cookie";
+
 export {};
 declare global {
   namespace Cypress {
     interface Chainable {
       login(username: string, password: string): Chainable<void>;
+      deleteShop(): Chainable<void>;
     }
   }
 }
@@ -56,5 +60,29 @@ Cypress.Commands.add('login', (username, password) => {
     cy.get('input[type=password]').type(password);
     cy.get('button[type=submit]').click();
     cy.url().should('contain', '/dashboard');
+  });
+});
+
+Cypress.Commands.add('deleteShop', () => {
+  cy.session([], () => {
+    const access_token = Cookies.get('access');
+    fetch(`${apiUrl}/shops/shops/me/`, {
+    method: 'GET',
+    headers: {
+      Authorization: `JWT ${access_token}`,
+      'Content-Type': 'application/json; charset=UTF-8',
+    }
+    }).then((response) => {
+      response.json().then((data) => {
+        console.log(data);
+        fetch(`${apiUrl}/shops/shops/${data.id}`, {
+          method: 'DELETE',
+          headers: {
+            Authorization: `JWT ${access_token}`,
+            'Content-Type': 'application/json; charset=UTF-8',
+          }
+        })
+      })
+    });
   });
 });
